Move interpreter
